requirements.txt
asgiref
boto3
Django
django-admin-sortable2
django-ckeditor
django-cleanup
django-cors-headers
django-filter
django-jazzmin
django-js-asset
djangorestframework
django-rest-registration
django-storages
drf-yasg
django-phonenumber-field
django-resized
gunicorn
phonenumbers
psycopg2-binary
python-decouple
Pillow
sqlparse